wave: Optionally preserve history when importing.
As a side-effect of this change, imported waves are no longer shown in
the inbox until they are edited.  This will fix itself when the
full-text-search branch is merged.  You can still find imported waves
via the links in /import , except if you import by id, in which case
you'll need to find the imported ids in server logs, or use the
full-text-search branch.

Also limit queue rate and concurrency by default, since for now, we
should assume that most installations will be small.