Create new backend import-worker and use it for all import tasks.
This allows import tasks to run for >10 minutes, useful when
preserving history for wavelets with very long histories.

Also drop slob backend for now since nobody uses it (as far as I know)
and it complicates the build script (not fundamentally, just in Ant
syntax).  Should resurrect this at some point but as dynamic backend
and with a smarter load balancing algorithm.