Started implementing part of the Passive Robot api.
The current code is able to generate events and send the JSON to the Robot. Robot response processing is next. There are a few other things to fix and do as well marked throughout the code but its promising to see that it is able to talk to my old robots :).